Finish separating core logic from implementation details
